requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
django-filter>=23.5
drf-spectacular>=0.27.0
google-generativeai>=0.3.0
celery>=5.3.0
redis>=5.0.0
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
